ai/smile-detection-ai/requirements.txt
# Deep Learning
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
tensorboard>=2.15.0
onnxruntime>=1.16.0

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8
albumentations==1.3.1
Pillow==10.1.0

# Data Processing
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# API Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0

# Utilities
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
pyyaml==6.0.1
mlflow>=2.9.0

# Dataset Download
kaggle==1.5.16
gdown==4.7.1
